+++ api_v3/lib/types/partner/KalturaPartner.php
@@ -440,6 +440,12 @@ class KalturaPartner extends KalturaObject implements IFilterable
 	public $numPrevPassToKeep;
 	
 	/**
+	 * @var bool
+	 * @readonly
+	 */
+	public $allowDefaultPasswordRestrictions;
+	
+	/**
 	 * @var KalturaTwoFactorAuthenticationMode
 	 * @readonly
 	 */
@@ -469,6 +475,12 @@ class KalturaPartner extends KalturaObject implements IFilterable
 	 */
 	public $eventPlatformAllowedTemplates;
 	
+	/**
+	 * @var int
+	 * @readonly
+	 */
+	public $verticalClassificationId;
+	
 	private static $map_between_objects = array
 	(
 		'id' , 'name', 'website' => 'url1' , 'notificationUrl' => 'url2' , 'appearInSearch' , 'createdAt' , 'adminName' , 'adminEmail' , 'blockDirectLogin',
@@ -481,7 +493,8 @@ class KalturaPartner extends KalturaObject implements IFilterable
 		'publisherEnvironmentType', 'ovpEnvironmentUrl', 'ottEnvironmentUrl', 'authenticationType', 'extendedFreeTrailExpiryReason', 'extendedFreeTrailExpiryDate',
 		'extendedFreeTrail', 'extendedFreeTrailEndsWarning', 'eightyPercentWarning', 'usageLimitWarning', 'lastFreeTrialNotificationDay','monitorUsage', 'additionalParams',
 		'passwordStructureValidations', 'passReplaceFreq', 'maxLoginAttempts', 'loginBlockPeriod', 'numPrevPassToKeep', 'twoFactorAuthenticationMode', 'isSelfServe', 'allowedDomains',
-		'excludedAdminRoleName', 'eventPlatformAllowedTemplates'
+		'excludedAdminRoleName', 'eventPlatformAllowedTemplates', 'verticalClassificationId' => 'verticalClasiffication', 'allowDefaultPasswordRestrictions'
+
 	);
 	
 	public function getMapBetweenObjects ( )
@@ -506,6 +519,7 @@ class KalturaPartner extends KalturaObject implements IFilterable
 		$this->adminName = kString::stripUtf8InvalidChars($this->adminName);
 		$this->describeYourself = kString::stripUtf8InvalidChars($this->describeYourself);
 		$this->additionalParams = KalturaKeyValueArray::fromKeyValueArray($partner->getAdditionalParams());
+		$this->allowDefaultPasswordRestrictions = !$partner->getEnabledService(PermissionName::FEATURE_DISABLE_PASSWORD_RESTRICTION);
 		if (!$this->host){
 			$this->host = null;
 		}